# migracao/contracts/openapi/internal-api.yaml
openapi: 3.1.0
info:
  title: Saltcorn Go — API interna (BFF → Go)
  version: "0.1.0"
  description: >
    Contrato consumido pelo BFF Node.js (ADR-0003) — nunca diretamente pelo
    navegador — e, em versões futuras, por integrações externas autorizadas
    via API pública própria. Cobre o exemplo de CQRS de registros dinâmicos
    definido em ADR-0001 (`CreateRecord`/`ListRecords`) como recorte inicial
    representativo: não é a superfície completa do domínio, que ainda não
    existe em Go (ver docs/migracao-go/execucoes/GO-006.md, nota de escopo).
servers:
  - url: http://localhost:8090
    description: cmd/server local (GO-005)
security:
  - ServiceIdentity: []
paths:
...
  /v1/tenants/{tenant}/actor:
    parameters:
      - name: tenant
        in: path
        required: true
        schema:
          $ref: "./common.yaml#/components/schemas/Tenant"
    get:
      operationId: getActor
      summary: "Query: papel atual do ator da identidade delegada"
      description: >
        Adicionado por GO-017 para o BFF resolver papel/permissões atuais a
        cada requisição (ADR-0007) em vez de confiar num valor cacheado na
        sessão de navegador — uma mudança de papel ou revogação de acesso
        vale a partir desta consulta, nunca de um valor antigo guardado
        alhures. `id`/`role_id` são resolvidos do `sub` do token, não de
        parâmetro algum — não é possível consultar o ator de outra pessoa.
      responses:
        "503":
          description: Capacidade temporariamente esgotada (service_unavailable); Retry-After em segundos
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
        "200":
          description: papel atual do ator autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
        "401":
          $ref: "./common.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./common.yaml#/components/responses/Forbidden"
  /v1/tenants/{tenant}/tables:
    parameters:
      - name: tenant
        in: path
        required: true
        schema:
          $ref: "./common.yaml#/components/schemas/Tenant"
    post:
      operationId: createTable
      summary: "Command: cria uma tabela dinâmica"
      description: >
        Adicionado por GO-019 — `internal/metadata.CreateTable` (GO-011)
        nunca tinha exposição HTTP. Sem Idempotency-Key: já é idempotente
        por definição própria (uma definição idêntica a uma já existente é
        um no-op bem-sucedido).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableInput"
      responses:
        "503":
          description: Capacidade temporariamente esgotada (service_unavailable); Retry-After em segundos
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
        "201":
          description: tabela criada (ou já existente, idempotente)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
        "401":
          $ref: "./common.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./common.yaml#/components/responses/Forbidden"
  /v1/tenants/{tenant}/tables/{table}/fields:
    parameters:
      - name: tenant
        in: path
        required: true
        schema:
          $ref: "./common.yaml#/components/schemas/Tenant"
      - name: table
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: addField
      summary: "Command: adiciona um campo a uma tabela dinâmica"
      description: Adicionado por GO-019 — `internal/metadata.AddField` (GO-011) nunca tinha exposição HTTP. Idempotente pela mesma razão de createTable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldInput"
      responses:
        "503":
          description: Capacidade temporariamente esgotada (service_unavailable); Retry-After em segundos
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
        "201":
          description: campo criado (ou já existente, idempotente)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Field"
        "401":
          $ref: "./common.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./common.yaml#/components/responses/Forbidden"
  /v1/tenants/{tenant}/views:
    parameters:
      - name: tenant
        in: path
        required: true
        schema:
          $ref: "./common.yaml#/components/schemas/Tenant"
    get:
      operationId: listViews
      summary: "Query: lista views"
      description: >
        Adicionado por GO-020 — a página administrativa de views (frontend)
        precisa enumerar o que existe antes de abrir uma view individual.
        `table` filtra por tabela; omitido lista todas as views visíveis ao
        ator (mesmo filtro por papel de getView/updateView — nunca revela
        uma view não publicada a um ator sem papel suficiente).
      parameters:
        - name: table
          in: query
          schema:
            type: string
      responses:
        "503":
          description: Capacidade temporariamente esgotada (service_unavailable); Retry-After em segundos
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
        "200":
          description: views visíveis ao ator
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/View"
        "401":
          $ref: "./common.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./common.yaml#/components/responses/Forbidden"
    post:
      operationId: createView
      summary: "Command: cria uma view (documento de layout do editor)"
      description: >
        Adicionado por GO-019 — persiste o documento que o editor Craft.js
        (GO-018) produz (`craftToSaltcorn`). Nasce com MinRole = admin
        (nunca publicada por omissão); publicar é um `updateView` baixando
        `min_role`. Idempotente por Idempotency-Key (ao contrário de
        createTable/addField, uma view não é idempotente por definição
        própria — duas criações com o mesmo nome sem isso colidiriam em
        `duplicate_name` num retry legítimo).
      parameters:
        - $ref: "./common.yaml#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViewInput"
      responses:
        "503":
          description: Capacidade temporariamente esgotada (service_unavailable); Retry-After em segundos
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
        "201":
          description: view criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/View"
        "401":
          $ref: "./common.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./common.yaml#/components/responses/Forbidden"
        "409":
          $ref: "./common.yaml#/components/responses/IdempotencyConflict"
  /v1/tenants/{tenant}/views/{id}:
    parameters:
      - name: tenant
        in: path
        required: true
        schema:
          $ref: "./common.yaml#/components/schemas/Tenant"
      - name: id
        in: path
        required: true
        schema:
          $ref: "./common.yaml#/components/schemas/Id"
    get:
      operationId: getView
      summary: "Query: reabre uma view"
      description: O papel do ator é resolvido de novo a cada chamada — uma view despublicada nega a leitura seguinte sem nenhum resquício da anterior (mesmo espírito de GO-015).
      responses:
        "503":
          description: Capacidade temporariamente esgotada (service_unavailable); Retry-After em segundos
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
        "200":
          description: view encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/View"
        "401":
          $ref: "./common.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./common.yaml#/components/responses/Forbidden"
        "404":
          description: view não encontrada
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
    patch:
      operationId: updateView
      summary: "Command: salva ou publica uma view"
      description: >
        A mesma operação serve para "salvar" (corpo com `configuration`
        nova) e "publicar" (corpo com `min_role` menor) — só campos
        diferentes do mesmo PATCH. Exige `_version` (o `_version` de uma
        leitura anterior) para controle de concorrência otimista: se a
        view mudou desde a leitura, 409 — "conflito de edição é
        apresentado sem sobrescrever silenciosamente" (critério de aceite
        de GO-019). Idempotente por Idempotency-Key. Desde GO-020, publicar
        (resultar num `min_role` que não seja admin-only) exige que o
        layout seja executável pelo runtime novo — 422 caso contrário
        ("layouts incompatíveis bloqueiam publicação", critério de aceite
        de GO-020).
      parameters:
        - $ref: "./common.yaml#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViewUpdateInput"
      responses:
        "503":
          description: Capacidade temporariamente esgotada (service_unavailable); Retry-After em segundos
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
        "200":
          description: view atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/View"
        "401":
          $ref: "./common.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./common.yaml#/components/responses/Forbidden"
        "409":
          description: conflito de versão otimista (edição concorrente) ou de idempotência
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
        "422":
          description: >
            layout incompatível com o runtime de renderização — só ao
            tentar publicar (min_role não admin-only); editar mantendo
            admin-only nunca é bloqueado.
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
              example:
                error:
                  code: view_unsupported
                  message: "template \"Feed\" não suportado neste runtime (só \"List\", \"Show\", \"Edit\")"
...
  /v1/tenants/{tenant}/views/{id}/rows/{recordId}:
    parameters:
      - name: tenant
        in: path
        required: true
        schema:
          $ref: "./common.yaml#/components/schemas/Tenant"
      - name: id
        in: path
        required: true
        schema:
          $ref: "./common.yaml#/components/schemas/Id"
      - name: recordId
        in: path
        required: true
        schema:
          $ref: "./common.yaml#/components/schemas/Id"
    delete:
      operationId: deleteViewRow
      summary: "Command: ação de coluna \"Delete\" de uma view List (GO-039)"
      description: >
        Só aceita um id de view cujo `configuration.columns` de fato
        declara uma coluna `Action`/`Delete` (422 caso contrário — nunca
        um atalho para deletar de qualquer view List). Confere o
        `minRole` do PRÓPRIO nó de coluna, além do `min_role_write` da
        tabela por baixo (internal/records) — duas checagens
        independentes. `?version=` é o mesmo `_version` de controle de
        concorrência otimista de qualquer outra escrita.
      parameters:
        - name: version
          in: query
          required: true
          schema:
            type: string
      responses:
        "503":
          description: Capacidade temporariamente esgotada (service_unavailable); Retry-After em segundos
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
        "204":
          description: linha excluída
        "400":
          description: "?version= ausente"
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
        "401":
          $ref: "./common.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./common.yaml#/components/responses/Forbidden"
        "404":
          description: view ou registro não encontrado
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
        "409":
          description: conflito de versão otimista
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
        "422":
          description: view existe, mas não declara uma ação de coluna "Delete"
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/Error"
  /v1/tenants/{tenant}/users:
    parameters:
      - name: tenant
        in: path
        required: true
        schema:
          $ref: "./common.yaml#/components/schemas/Tenant"
    get:
      operationId: listUsers
      summary: "Query: lista usuários do tenant (admin)"
      description: >
        Adicionado por GO-044 — a administração de usuários do legado
        (`auth/admin.ts`) que faltava expor. Só um ator admin pode listar
        (`identity.ListUsers`); um ator sem papel suficiente recebe 403.
      responses:
        "200":
          description: usuários do tenant, em ordem de criação
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "./common.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./common.yaml#/components/responses/Forbidden"
...
components:
  securitySchemes:
    ServiceIdentity:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Identidade delegada ator+tenant entre BFF e backend Go (ADR-0003). O
        BFF assina um JWT de vida curta com claims `{sub (ator), tenant, iat,
        exp}`; o Go valida assinatura, expiração, e que o claim `tenant` bate
        com o tenant da URL (ver schema Tenant em common.yaml). Nunca aceitar
        tenant vindo de um header solto — GO-001 §2.1, "identidade não pode
        ser forjada por headers". Exemplos completos (positivo e dois
        negativos) em ../fixtures/identidade-delegada.md.
  schemas:
    Actor:
      type: object
      required: [id, role_id]
      properties:
        id:
          $ref: "./common.yaml#/components/schemas/Id"
        role_id:
          type: integer
          description: Papel atual do ator (identity.RoleID) — nunca cacheado pelo chamador, sempre resolvido nesta consulta.
...